requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
crispy-bootstrap4==2024.10
cryptography==41.0.3
defusedxml==0.7.1
diff-match-patch==20241021
dj-database-url==2.3.0
Django==5.0.2
django-allauth==0.54.0
django-bootstrap-icons==0.9.0
django-bootstrap-v5==1.0.11
django-crispy-forms==2.3
django-etc==1.4.0
django-extensions==3.2.3
django-filter==24.3
django-heroku==0.3.1
django-hitcount==1.3.5
django-import-export==4.3.3
django-jazzmin==2.6.0
django-js-asset==3.0.1
django-resized==1.0.3
django-suit==0.2.28
django-taggit==6.1.0
django-tinymce==4.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
Faker==33.0.0
idna==3.4
jwt==1.3.1
Markdown==3.7
numpy==2.2.1
oauthlib==3.2.2
pillow==11.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.6
sqlparse==0.4.4
tablib==3.7.0
typing_extensions==4.7.1
tzdata==2023.4
urllib3==2.0.4
whitenoise==6.8.2
